实现标头字段中的“长度模式”位的判断和对应的功能
位运算好累www

+++ library/bintp1.h
@@ -10,9 +10,9 @@
     Size is the number of bytes
  */
 struct BintpFieldPair {
-    uint16_t name_size; // max is ~(-32768) == 32768, don't use int16
+    size_t name_size;
     void *name;
-    uint16_t value_size;
+    size_t value_size;
     void *value;
 };
